backend/requirements.txt
fastapi
mangum
python-jose
boto3
requests
uvicorn